extracts email verification into its own recipe
