+++ build.gradle
@@ -2,7 +2,7 @@ plugins {
     id 'java-library'
 }
 
-version = "2.5.0"
+version = "2.6.0"
 
 repositories {
     jcenter()
